feat: added feature screening, chat bot

+++ app/src/main/java/com/manikandareas/codileap/user/data/networking/dto/GetUserProgressResponseDto.kt
@@ -24,5 +24,5 @@ data class OverallProgressDto(
 data class LearningPathProgressDto(
     val pathId: Int,
     val pathName: String,
-    val progress: Int
+    val progress: Float
 )
